+++ packages/playground/vite.config.ts
@@ -1,15 +1,7 @@
 import { defineConfig } from 'vite'
-import VitePluginCertificate from 'vite-plugin-mkcert'
+import mkcert from 'vite-plugin-mkcert'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  server: {
-    open: false,
-    https: true
-  },
-  plugins: [
-    VitePluginCertificate({
-      source: 'coding'
-    })
-  ]
+  plugins: [mkcert()]
 })